refactor: resolvers
