+++ src/app/lib/graphql/resolvers/country.ts
@@ -1,0 +1,15 @@
+import { Resolvers } from "@apollo/client";
+import { Context } from "@app/api/graphql/route";
+export default {
+  Query: {
+    countries: async (
+      _parent: undefined,
+      args: undefined,
+      context: Context
+    ) => {
+      return await context.prisma.country.findMany({
+        orderBy: { name: "asc" },
+      });
+    },
+  },
+} as Resolvers;
